feat(balance): KR/US 잔고 페이지를 토글 단일 페이지로 통합
- /balance 를 KR/US 전환 토글이 있는 단일 페이지로 변경 (기존 선택 화면 대체)
- balance-kr / balance-us 페이지 본문을 components/balance/balanceKrView, balanceUsView 로 추출
- 기존 /balance-kr, /balance-us 는 ?key= 보존하며 /balance?country= 로 리다이렉트
- 뷰의 router.replace 가 country 등 기존 쿼리를 보존하도록 수정
- middleware admin 가드를 /balance 전체로 확장
