+++ cf-idiotquant/app/(balance)/balance/page.tsx
@@ -1,69 +1,73 @@
-"use client"
+"use client";
 
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { useSession } from "next-auth/react";
-import { selectKakaoTotal } from "@/lib/features/kakao/kakaoSlice";
-import { useAppSelector } from "@/lib/hooks";
-import SpotlightCard from "@/src/Components/SpotlightCard/SpotlightCard";
-import { LockClosedIcon, LockOpenIcon, WalletIcon } from "@heroicons/react/24/outline";
-import Link from "next/link";
+import { Suspense, useCallback } from "react";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { LoadingState } from "@/components/balance/shared";
+import { BalanceKrView } from "@/components/balance/balanceKrView";
+import { BalanceUsView } from "@/components/balance/balanceUsView";
+import { cn } from "@/lib/utils";
 
-const DEBUG = false;
+type Country = "kr" | "us";
 
-export default function Balance() {
-    const router = useRouter();
-    const { data: session, status } = useSession();
-    const kakaoTotal = useAppSelector(selectKakaoTotal);
+// =========================================================================
+// KR / US 전환 토글
+// =========================================================================
+function CountryToggle({ country, onChange }: { country: Country; onChange: (c: Country) => void }) {
+  const items: { id: Country; label: string; flag: string }[] = [
+    { id: "kr", label: "한국", flag: "🇰🇷" },
+    { id: "us", label: "미국", flag: "🇺🇸" },
+  ];
+  return (
+    <div className="flex items-center gap-1 bg-[#faf9f7] dark:bg-[#35332e] p-1 rounded-xl">
+      {items.map(it => (
+        <button
+          key={it.id}
+          onClick={() => onChange(it.id)}
+          className={cn(
+            "flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-black transition-all whitespace-nowrap",
+            country === it.id
+              ? "bg-white dark:bg-[#1a1915] text-neutral-950 dark:text-white shadow-sm"
+              : "text-neutral-500 hover:text-neutral-800 dark:hover:text-neutral-300"
+          )}
+        >
+          <span aria-hidden>{it.flag}</span>
+          {it.label}
+        </button>
+      ))}
+    </div>
+  );
+}
 
-    useEffect(() => {
-        if (status === "loading") return;
-        if (!session || session.user?.name !== process.env.NEXT_PUBLIC_MASTER) {
-            router.replace("/");
-        }
-    }, [session, status, router]);
+// =========================================================================
+// 통합 페이지 본문
+// =========================================================================
+function BalancePage() {
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const country: Country = searchParams.get("country") === "us" ? "us" : "kr";
 
-    if (status === "loading" || !session || session.user?.name !== process.env.NEXT_PUBLIC_MASTER) {
-        return null;
-    }
+  const handleChange = useCallback((next: Country) => {
+    if (next === country) return;
+    const params = new URLSearchParams(searchParams.toString());
+    params.set("country", next);
+    router.replace(`${pathname}?${params.toString()}`);
+  }, [country, pathname, router, searchParams]);
 
-    return (
-        <>
-            <div className="flex flex-col dark:!bg-black h-full dark:!text-white">
-                <div className="text-2xl flex items-center justify-left gap-2 p-6 sm:p-8 md:p-10 lg:p-12 ">
-                    <WalletIcon className="h-5 w-5" strokeWidth={2} />
-                    <div>account inquiry</div>
-                    {!!!kakaoTotal?.id ? <LockClosedIcon className="h-5 w-5" strokeWidth={2} /> : <LockOpenIcon className="h-5 w-5" strokeWidth={2} />}
-                </div>
-                <div className="grid grid-cols-2 items-center justify-center h-full p-6 gap-6 sm:p-8 sm:gap-8 md:p-10 md:gap-10 lg:p-12 lg:gap-12">
-                    <Link href="/balance-kr">
-                        <SpotlightCard className="!bg-white dark:!bg-black dark:!border dark:!border-white shadow-md" spotlightColor="rgba(60, 130, 246, 0.5)">
-                            <div className="flex flex-col items-center justify-center text-black dark:!text-white">
-                                <div className="text-2xl">
-                                    Korea
-                                </div>
-                                <div className="text-xs sm:!text-sm md:!text-base lg:!text-lg xl:!text-xl 2xl:!text-2xl">
-                                    (KOSPI/KOSDAQ/KONEX)
-                                </div>
-                            </div>
-                        </SpotlightCard>
-                    </Link>
-                    <Link href="/balance-us">
-                        <SpotlightCard className="!bg-white dark:!bg-black dark:!border dark:!border-white shadow-md" spotlightColor="rgba(60, 130, 246, 0.5)">
-                            <div className="flex flex-col items-center justify-center text-black dark:!text-white">
-                                <div className="text-2xl">
-                                    US
-                                </div>
-                                <div className="text-xs sm:!text-sm md:!text-base lg:!text-lg xl:!text-xl 2xl:!text-2xl">
-                                    (NASDAQ)
-                                </div>
-                            </div>
-                        </SpotlightCard>
-                    </Link>
-                </div>
-            </div>
-            <div className="dark:!bg-black h-lvh"></div>
+  const toggle = <CountryToggle country={country} onChange={handleChange} />;
 
-        </>
-    )
+  return country === "us"
+    ? <BalanceUsView countryToggle={toggle} />
+    : <BalanceKrView countryToggle={toggle} />;
+}
+
+// =========================================================================
+// 페이지 내보내기
+// =========================================================================
+export default function Page() {
+  return (
+    <Suspense fallback={<LoadingState message="계좌 데이터를 불러오는 중..." />}>
+      <BalancePage />
+    </Suspense>
+  );
 }
